ajax login

+++ service/src/main/java/com/tipcrm/service/WebContext.java
@@ -1,7 +1,7 @@
 package com.tipcrm.service;
 public interface WebContext {
 
-    Long getCurrentUserId();
+    Integer getCurrentUserId();
 
     String getCurrentUserName();
 }
